+++ FeaturesProvider/.env
@@ -1,0 +1,1 @@
+API_KEY=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
