+++ FeaturesProvider/.gitignore
@@ -1,5 +1,4 @@
 # Environments
-.env
 .venv
 env/
 venv/
