+++ WebInterface/src/components/navigation.tsx
