+++ WebInterface/src/vite-env.d.ts
