--- WebInterface/vite-env.d.ts
@@ -1,8 +1,0 @@
-interface ImportMetaEnv {
-    readonly VITE_API_HOST?: string;
-    readonly VITE_API_PORT?: string;
-}
-
-interface ImportMeta {
-    readonly env: ImportMetaEnv;
-}
