+++ WebInterface/vite.config.ts
@@ -1,11 +1,15 @@
-import {defineConfig} from 'vite'
-import react from '@vitejs/plugin-react'
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
+import path from 'path';
 
-// https://vite.dev/config/
 export default defineConfig({
-    plugins: [react()],
-    server: {
-        host: process.env.VITE_DEV_HOST || 'localhost',  // Default to localhost
-        port: Number(process.env.VITE_DEV_PORT) || 3000, // Default to 3000
+  plugins: [react()],
+  optimizeDeps: {
+    exclude: ['lucide-react'],
+  },
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname, './src'),
     },
-})
+  },
+});
