+++ logs/feat.out
@@ -1,0 +1,22 @@
+Activating virtual environment and installing dependencies...
+Requirement already satisfied: fastapi in ./venv/lib/python3.11/site-packages (from -r requirements.txt (line 1)) (0.115.12)
+Requirement already satisfied: uvicorn in ./venv/lib/python3.11/site-packages (from -r requirements.txt (line 2)) (0.34.0)
+Requirement already satisfied: httpx in ./venv/lib/python3.11/site-packages (from -r requirements.txt (line 3)) (0.28.1)
+Requirement already satisfied: requests in ./venv/lib/python3.11/site-packages (from -r requirements.txt (line 4)) (2.32.3)
+Requirement already satisfied: python-dotenv in ./venv/lib/python3.11/site-packages (from -r requirements.txt (line 5)) (1.1.0)
+Requirement already satisfied: starlette<0.47.0,>=0.40.0 in ./venv/lib/python3.11/site-packages (from fastapi->-r requirements.txt (line 1)) (0.46.1)
+Requirement already satisfied: pydantic!=1.8,!=1.8.1,!=2.0.0,!=2.0.1,!=2.1.0,<3.0.0,>=1.7.4 in ./venv/lib/python3.11/site-packages (from fastapi->-r requirements.txt (line 1)) (2.11.3)
+Requirement already satisfied: typing-extensions>=4.8.0 in ./venv/lib/python3.11/site-packages (from fastapi->-r requirements.txt (line 1)) (4.13.1)
+Requirement already satisfied: click>=7.0 in ./venv/lib/python3.11/site-packages (from uvicorn->-r requirements.txt (line 2)) (8.1.8)
+Requirement already satisfied: h11>=0.8 in ./venv/lib/python3.11/site-packages (from uvicorn->-r requirements.txt (line 2)) (0.14.0)
+Requirement already satisfied: anyio in ./venv/lib/python3.11/site-packages (from httpx->-r requirements.txt (line 3)) (4.9.0)
+Requirement already satisfied: certifi in ./venv/lib/python3.11/site-packages (from httpx->-r requirements.txt (line 3)) (2025.1.31)
+Requirement already satisfied: httpcore==1.* in ./venv/lib/python3.11/site-packages (from httpx->-r requirements.txt (line 3)) (1.0.7)
+Requirement already satisfied: idna in ./venv/lib/python3.11/site-packages (from httpx->-r requirements.txt (line 3)) (3.10)
+Requirement already satisfied: charset-normalizer<4,>=2 in ./venv/lib/python3.11/site-packages (from requests->-r requirements.txt (line 4)) (3.4.1)
+Requirement already satisfied: urllib3<3,>=1.21.1 in ./venv/lib/python3.11/site-packages (from requests->-r requirements.txt (line 4)) (2.3.0)
+Requirement already satisfied: annotated-types>=0.6.0 in ./venv/lib/python3.11/site-packages (from pydantic!=1.8,!=1.8.1,!=2.0.0,!=2.0.1,!=2.1.0,<3.0.0,>=1.7.4->fastapi->-r requirements.txt (line 1)) (0.7.0)
+Requirement already satisfied: pydantic-core==2.33.1 in ./venv/lib/python3.11/site-packages (from pydantic!=1.8,!=1.8.1,!=2.0.0,!=2.0.1,!=2.1.0,<3.0.0,>=1.7.4->fastapi->-r requirements.txt (line 1)) (2.33.1)
+Requirement already satisfied: typing-inspection>=0.4.0 in ./venv/lib/python3.11/site-packages (from pydantic!=1.8,!=1.8.1,!=2.0.0,!=2.0.1,!=2.1.0,<3.0.0,>=1.7.4->fastapi->-r requirements.txt (line 1)) (0.4.0)
+Requirement already satisfied: sniffio>=1.1 in ./venv/lib/python3.11/site-packages (from anyio->httpx->-r requirements.txt (line 3)) (1.3.1)
+Starting the Features Provider API...
