+++ logs/web.out
@@ -1,0 +1,9 @@
+
+> trackmyoffer@0.0.0 dev
+> vite
+
+
+  VITE v5.4.17  ready in 83 ms
+
+  ➜  Local:   http://localhost:5173/
+  ➜  Network: use --host to expose
