--- project/.env.example
@@ -1,2 +1,0 @@
-VITE_API_HOST=localhost
-VITE_API_PORT=8080
